Moved the base wrappers from DockableWindowHost into their own namespace.
